requirements.txt
web3~=7.4.0
python-dotenv==1.0.1
tabulate==0.9.0
autopep8==2.3.1